Extract Options classes into separate files under Options/
Move each GlobalOptions subclass from its command file into a
dedicated file in DataFactory.MCP.Fabric/Options/ for cleaner
separation of concerns.
