+++ DataFactory.MCP.Fabric/Commands/GetPipelineCommand.cs
@@ -72,9 +72,3 @@ public sealed class GetPipelineCommand(
         return context.Response;
     }
 }
-
-public sealed class GetPipelineOptions : GlobalOptions
-{
-    public string WorkspaceId { get; set; } = string.Empty;
-    public string PipelineId { get; set; } = string.Empty;
-}
