+++ DataFactory.MCP.Fabric/Options/CreatePipelineOptions.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using Microsoft.Mcp.Core.Options;
+
+namespace DataFactory.MCP.Fabric.Options;
+
+public sealed class CreatePipelineOptions : GlobalOptions
+{
+    public string WorkspaceId { get; set; } = string.Empty;
+    public string DisplayName { get; set; } = string.Empty;
+    public string? Description { get; set; }
+}
