+++ DataFactory.MCP.Fabric/Options/GetPipelineOptions.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using Microsoft.Mcp.Core.Options;
+
+namespace DataFactory.MCP.Fabric.Options;
+
+public sealed class GetPipelineOptions : GlobalOptions
+{
+    public string WorkspaceId { get; set; } = string.Empty;
+    public string PipelineId { get; set; } = string.Empty;
+}
