+++ DataFactory.MCP.Fabric/Options/ListWorkspacesOptions.cs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using Microsoft.Mcp.Core.Options;
+
+namespace DataFactory.MCP.Fabric.Options;
+
+public sealed class ListWorkspacesOptions : GlobalOptions
+{
+    public string? Roles { get; set; }
+}
